+++ Source/Components/Touchpoints/MLBusinessTouchpointsRegistry.swift
@@ -8,25 +8,27 @@
 import Foundation
 
 protocol MLBusinessTouchpointsRegistryProtocol {
-    func mapper(for type: String) -> MLBusinessTouchpointsMapperProtocol?
+    func mapper(for type: String) -> MLBusinessMapperProtocol?
     func views(for type: String) -> MLBusinessTouchpointsBaseView.Type?
 }
 
 class MLBusinessTouchpointsRegistry: MLBusinessTouchpointsRegistryProtocol {
-    private var mappers: [String: MLBusinessTouchpointsMapperProtocol] = [:]
+    private var mappers: [String: MLBusinessMapperProtocol] = [:]
     private var views: [String: MLBusinessTouchpointsBaseView.Type] = [:]
 
     init() {
-        register(type: "GRID", mapper: MLBusinessTouchpointsMapper<MLBusinessTouchpointsGridModel>(), view: MLBusinessTouchpointsGridView.self)
-        register(type: "CAROUSEL", mapper: MLBusinessTouchpointsMapper<MLBusinessTouchpointsCarouselModel>(), view: MLBusinessTouchpointsCarouselView.self)
+        register(type: "GRID", mapper: MLBusinessMapper<MLBusinessTouchpointsGridModel>(), view: MLBusinessTouchpointsGridView.self)
+        register(type: "CAROUSEL", mapper: MLBusinessMapper<MLBusinessTouchpointsCarouselModel>(), view: MLBusinessTouchpointsCarouselView.self)
+        register(type: "HYBRID_CAROUSEL", mapper: MLBusinessMapper<MLBusinessHybridCarouselModel>(), view: MLBusinessHybridCarouselView.self)
+        register(type: "MULTIPLE_ROW", mapper: MLBusinessMapper<MLBusinessTouchpointsMultipleRowModel>(), view: MLBusinessTouchpointsMultipleRowView.self)
     }
 
-    func register(type: String, mapper: MLBusinessTouchpointsMapperProtocol, view: MLBusinessTouchpointsBaseView.Type) {
+    func register(type: String, mapper: MLBusinessMapperProtocol, view: MLBusinessTouchpointsBaseView.Type) {
         views[type] = view
         mappers[type] = mapper
     }
 
-    func mapper(for type: String) -> MLBusinessTouchpointsMapperProtocol? {
+    func mapper(for type: String) -> MLBusinessMapperProtocol? {
         return mappers[type]
     }