+++ Source/Components/Touchpoints/Types/HybridCarousel/Card/MLBusinessHybridCarouselCardTypes.swift
@@ -1,0 +1,13 @@
+//
+//  MLBusinessHybridCarouselCardTypes.swift
+//  MLBusinessComponents
+//
+//  Created by Vicente Veltri on 04/08/2020.
+//
+
+import Foundation
+
+struct MLBusinessHybridCarouselCardTypes {
+    public static let defaultType = "DEFAULT"
+    public static let viewMoreType = "VIEW_MORE"
+}