+++ Source/Components/Touchpoints/Types/MultipleRow/MLBusinessTouchpointsMultipleRowData.swift
@@ -1,0 +1,15 @@
+//
+//  MLBusinessTouchpointsMultipleRowData.swift
+//  MLBusinessComponents
+//
+//  Created by Vicente Veltri on 12/08/2020.
+//
+
+import Foundation
+
+@objc public protocol MLBusinessTouchpointsMultipleRowData: NSObjectProtocol {
+    @objc optional func getTitle() -> String?
+    @objc optional func getSubtitle() -> String?
+    @objc func getItems() -> [MLBusinessMultipleRowItemModel]
+    @objc optional func getDiscountTracker() -> MLBusinessDiscountTrackerProtocol?
+}